+++ main/include/DepthMapDataset.hpp
@@ -17,9 +17,9 @@
 
 #pragma once
 
-#include <vector>
-#include <memory>
 #include <math/geometry.hpp>
+#include <memory>
+#include <vector>
 
 namespace tests
 {
@@ -57,8 +57,8 @@ class DepthMapDataset
 
   private:
     size_t size_;
-    fusion::math::Mat4d intrinsics_;
-    std::vector<fusion::math::Mat4d> poses_;
+    fusion::math::Mat4f intrinsics_;
+    std::vector<fusion::math::Mat4f> poses_;
     std::vector<std::unique_ptr<uint16_t>> depthImages_;
     std::vector<std::unique_ptr<uint8_t>> colorImages_;
     std::vector<FrameInfo> depthFramesInfo_;
